src/requirements.txt
asgiref==3.2.3
Babel==2.8.0
boto3==1.13.26
botocore==1.16.26
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.3
django-debug-toolbar==2.2
django-extensions==2.2.9
django-filter==2.2.0
django-guardian==2.2.0
django-phonenumber-field==4.0.0
django-recaptcha==2.0.6
django-storages==1.9.1
django-widget-tweaks==1.4.8
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
jmespath==0.10.0
phonenumbers==8.12.5
Pillow==7.0.0
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.0
urllib3==1.25.9
whitenoise==5.0.1
